refactor(stage4): extract MultiPeerHeadHarness, port Suite to consume it
