+++ integration/src/test/scala/hydrozoa/integration/stage4/Suite.scala
@@ -1,11 +1,8 @@
 package hydrozoa.integration.stage4
 
 import cats.data.ReaderT
-import cats.effect.{Deferred, IO, Ref, Resource}
+import cats.effect.{IO, Ref, Resource}
 import cats.implicits.*
-import com.comcast.ip4s.{Port, host}
-import com.suprnation.actor.event.Error as ActorError
-import com.suprnation.actor.{ActorContext, ActorSystem}
 import hydrozoa.config.head.coil.{CoilPeerData, CoilPeers}
 import hydrozoa.config.head.initialization.{InitializationParametersGenTopDown, generateInitialBlock}
 import hydrozoa.config.head.multisig.timing.TxTiming.BlockTimes.BlockCreationEndTime
@@ -14,84 +11,31 @@ import hydrozoa.config.head.network.CardanoNetwork
 import hydrozoa.config.head.parameters.generateHeadParameters
 import hydrozoa.config.head.{InitParamsType, generateHeadConfig, generateHeadConfigBootstrap}
 import hydrozoa.config.node.{MultiNodeConfig, NodeConfig}
+import hydrozoa.integration.harness.MultiPeerHeadHarness
+import hydrozoa.integration.harness.MultiPeerHeadHarness.{BackendMode, TransportMode}
 import hydrozoa.integration.stage4.EffectsLanded.BlockExpectation
 import hydrozoa.integration.stage4.Model.*
 import hydrozoa.lib.cardano.scalus.QuantizedTime.given_Ordering_QuantizedInstant.mkOrderingOps
 import hydrozoa.lib.cardano.scalus.QuantizedTime.quantize
 import hydrozoa.lib.logging.{ContraTracer, Slf4jMsg, Slf4jMsgFormat, Slf4jTracer, info, warn}
-import hydrozoa.multisig.backend.cardano.{CardanoBackend, CardanoBackendMock, MockState, yaciTestSauceGenesis}
-import hydrozoa.multisig.consensus.CardanoLiaison
-import hydrozoa.multisig.consensus.peer.{CoilPeerNumber, HeadPeerId, HeadPeerNumber, PeerId, PeerWallet}
-import hydrozoa.multisig.consensus.transport.{CoilPeerWsTransport, CoilPeerWsTransportEventFormat, CoilTransport, HubTransport, HubWsTransport, HubWsTransportEventFormat, InProcessHubCoilTransport, InProcessPeerTransport, NodeWsServer, NodeWsServerEventFormat, PeerTransport, PeerTransportEventFormat, WsPeerTransport}
+import hydrozoa.multisig.backend.cardano.yaciTestSauceGenesis
+import hydrozoa.multisig.consensus.peer.{CoilPeerNumber, HeadPeerNumber, PeerId, PeerWallet}
 import hydrozoa.multisig.ledger.block.{BlockBrief, BlockNumber}
-import hydrozoa.multisig.ledger.eutxol2.{EutxoL2Ledger, toUtxos}
+import hydrozoa.multisig.ledger.eutxol2.toUtxos
 import hydrozoa.multisig.ledger.event.RequestId.ValidityFlag
 import hydrozoa.multisig.ledger.event.{RequestId, RequestNumber}
 import hydrozoa.multisig.ledger.stack.{PartitionEffects, Stack, StackEffects}
-import hydrozoa.multisig.persistence.rocksdb.RocksDbBackendStore
-import hydrozoa.multisig.persistence.{BackendStore, Cf, InMemoryBackendStore, Persistence, PersistenceEvent, PersistenceEventFormat}
-import hydrozoa.multisig.{CoilMultisigRegimeManager, HeadMultisigRegimeManager, HeadMultisigRegimeManagerEvent, HeadMultisigRegimeManagerEventFormat}
-import java.nio.file.{Files, Path}
-import java.util.concurrent.TimeUnit
-import org.http4s.Uri
-import org.http4s.client.websocket.WSClient
-import org.http4s.jdkhttpclient.JdkWSClient
-import org.http4s.server.websocket.WebSocketBuilder2
+import hydrozoa.multisig.persistence.{BackendStore, Cf}
 import org.scalacheck.commands.{AnyCommand, ModelBasedSuite, ScenarioGen}
 import org.scalacheck.{Gen, Prop, PropertyM}
-import scala.concurrent.duration.{DurationInt, DurationLong, FiniteDuration}
+import scala.concurrent.duration.{DurationInt, FiniteDuration}
 import scalus.cardano.address.ShelleyAddress
-import scalus.cardano.ledger.rules.{Context, UtxoEnv}
-import scalus.cardano.ledger.{CertState, TransactionHash, TransactionInput, Utxos}
+import scalus.cardano.ledger.{TransactionHash, TransactionInput, Utxos}
 import test.{SeedPhrase, TestPeers, given}
 
 // ===================================
 // Stage 4 suite
 // ===================================
-
-/** Shared state assembled in `sutResource` between starting the `ActorSystem` and building the
-  * per-peer actor stacks: the mock L1 backend, the per-peer `Deferred`/`Ref` plumbing, and the
-  * settlement/coverage signals + targets consumed by the property runners.
-  *
-  *   - `cardanoBackend` — single mock L1 shared by every peer.
-  *   - `pendingConnsMap(p)` — completed in `acquireSut` with peer `p`'s full `Connections`.
-  *   - `blockBriefsMap(p)` / `stacksMap(p)` — capture sinks: every `BlockBrief.Intermediate` from
-  *     JL and every `Stack.HardConfirmed` from SCA, per peer.
-  *   - `submittedRequestIds` — every `RequestId` the model submits across the run.
-  *   - `fastSettlementSignal` / `fastSettlementTarget` — fires once every targeted request id has
-  *     surfaced in some peer's captured briefs; target is set by `beforeFinalize`.
-  *   - `slowCoverageSignal` / `slowCoverageTarget` — fires once every targeted block number is
-  *     covered by some hard-confirmed stack on *every* peer (cross-peer barrier).
-  */
-private case class PostSystemState(
-    cardanoBackend: CardanoBackend[IO],
-    pendingConnsMap: Map[HeadPeerNumber, Deferred[IO, HeadMultisigRegimeManager.Connections]],
-    blockBriefsMap: Map[HeadPeerNumber, Ref[IO, Vector[BlockBrief.Intermediate]]],
-    stacksMap: Map[HeadPeerNumber, Ref[IO, Vector[Stack.HardConfirmed]]],
-    coilStacksMap: Map[CoilPeerNumber, Ref[IO, Vector[Stack.HardConfirmed]]],
-    submittedRequestIds: Ref[IO, Vector[RequestId]],
-    fastSettlementSignal: Deferred[IO, Unit],
-    slowCoverageSignal: Deferred[IO, Unit],
-    fastSettlementTarget: Deferred[IO, Set[RequestId]],
-    slowCoverageTarget: Deferred[IO, Set[Int]],
-    effectsLanded: Ref[IO, Set[TransactionHash]],
-    effectsLandedSignal: Deferred[IO, Unit],
-    effectsLandedTarget: Deferred[IO, List[BlockExpectation]],
-    fallbackEnteredSignal: Deferred[IO, TransactionHash],
-)
-
-/** Selects the persistence backend for stage4 peers.
-  *
-  *   - [[InMemory]] (default) — per-peer [[InMemoryBackendStore]]. No disk I/O; test isolation is
-  *     automatic.
-  *   - [[RocksDb]] — per-peer RocksDB directory under `root`. Each peer gets `root/peer-N/`. The
-  *     default `root` is a fresh tempdir per `RocksDb()` invocation so concurrent runs don't
-  *     collide; pass an explicit path to keep the store around (e.g. for inspection). Use this when
-  *     reproducing on-disk compaction / batching behavior.
-  */
-enum BackendMode:
-    case InMemory
-    case RocksDb(root: Path = Files.createTempDirectory("stage4-rocksdb-"))
 
 case class Stage4Suite(
     label: String = "stage4",
@@ -145,691 +89,130 @@ case class Stage4Suite(
 
     override def canStartupNewSut(): Boolean = true
 
-    // Resources acquired:
-    //   1. `ActorSystem[IO]
-    //   2. Per-head-peer `BackendStore` (RocksDB on disk or in-memory)
-    //   3. `WsNetwork` via `transportSetup` — `Resource.pure` in Direct mode; in WebSocket, the web socket servers
-    //   4. Per-coil-peer `InMemoryBackendStore` (always in-memory).
-    //   5. `Stage4Sut` via `Resource.make` — release cancels the per-peer CardanoLiaison
-    //      polling fibers and the event-stream error drainer.
-    override def sutResource(state: ModelState): Resource[IO, Stage4Sut] = {
+    // Stage4 owns its own capture-observer Refs/Deferreds + handle extraction; the harness
+    // (MultiPeerHeadHarness) owns the ActorSystem, mock CardanoBackend, transports, per-peer
+    // HMRM with BackendStore + Persistence, per-coil CMRM, connections drain, error drainer,
+    // and CL tick fibers. Hooks thread the stage4-side tracers into each MRM.
+    override def sutResource(state: ModelState): Resource[IO, Stage4Sut] =
         val multiNodeConfig = state.params.multiNodeConfig
-        val cardanoInfo = multiNodeConfig.headConfig.cardanoInfo
-        val peers = multiNodeConfig.nodeConfigs.keys.toSeq.sortBy(p => p: Int)
-        val coilConfigs = state.params.coilNodeConfigs
+        val peers           = multiNodeConfig.nodeConfigs.keys.toSeq.sortBy(p => p: Int)
+        val coilConfigs     = state.params.coilNodeConfigs
+        val startEpochMs    = state.currentModelTime.getEpochSecond * 1000L
 
-        // Advance simulated clock to the head's start epoch BEFORE creating the ActorSystem.
-        // With TestControl, IO.sleep advances the virtual clock only while no actor fibers
-        // exist; once actors are started their ping loops compete with tickOne, so the sleep
-        // must come first (same pattern as stage1 Suite).
-        //
-        // In WS mode (real clock) we skip this: the configured start epoch is potentially
-        // years in the future, sleeping would block the test. Block production almost
-        // certainly won't fire under WS+real-clock until the model anchors startTime to
-        // `Instant.now() + small`. v1 of WS mode validates the transport layer; full property
-        // validation under WS is a follow-up.
-        val startEpochMs = state.currentModelTime.getEpochSecond * 1000L
-
-        // ------ Pre-system IO: clock alignment. ------
-        val preSystem: IO[Unit] = for {
-            // TestControl branch: jump the virtual clock from 0 to the head's start epoch.
-            // Without TestControl this would be a literal multi-decade real sleep, so it must
-            // be gated on `useTestControl`. The non-TestControl analogue is the
-            // `state.takeoffTime` wait below — `genInitialState` anchors `currentModelTime`
-            // at `now + 60s` for that mode, so wall-clock sleeping until the anchor is the
-            // right move.
-            _ <- IO.whenA(useTestControl)(
-              IO.sleep(FiniteDuration(startEpochMs, TimeUnit.MILLISECONDS))
-            )
-            // Non-TestControl branch: wait until the wall clock reaches `takeoffTime` so the
-            // model clock and the SUT wall clock coincide at command 1. Abort if setup
-            // overran the budget — better a loud failure than a test that starts with
-            // already-violated timing. Same shape as stage 1.
-            _ <- state.takeoffTime match {
-                case None => IO.unit
-                case Some(t) =>
-                    IO.realTimeInstant.flatMap { now =>
-                        if now.isAfter(t) then
-                            IO.raiseError(
-                              RuntimeException(
-                                "Stage4 sutResource: initialization took too long " +
-                                    s"(takeoff: $t, now: $now)"
+        for
+            blockBriefsMap <- Resource.eval(
+                                  peers.toList
+                                      .traverse(p =>
+                                          Ref[IO]
+                                              .of(Vector.empty[BlockBrief.Intermediate])
+                                              .map(p -> _)
+                                      )
+                                      .map(_.toMap)
+                              )
+            stacksMap <- Resource.eval(
+                             peers.toList
+                                 .traverse(p =>
+                                     Ref[IO].of(Vector.empty[Stack.HardConfirmed]).map(p -> _)
+                                 )
+                                 .map(_.toMap)
+                         )
+            coilStacksMap <- Resource.eval(
+                                 coilConfigs
+                                     .traverse { cc =>
+                                         val cn = cc.ownPeerId match
+                                             case PeerId.Coil(n) => n
+                                             case PeerId.Head(_) =>
+                                                 throw new IllegalStateException(
+                                                   "coil node config carries a head peer id"
+                                                 )
+                                         Ref[IO].of(Vector.empty[Stack.HardConfirmed]).map(cn -> _)
+                                     }
+                                     .map(_.toMap)
+                             )
+            submittedRequestIds   <- Resource.eval(Ref[IO].of(Vector.empty[RequestId]))
+            fastSettlementSignal  <- Resource.eval(IO.deferred[Unit])
+            slowCoverageSignal    <- Resource.eval(IO.deferred[Unit])
+            fastSettlementTarget  <- Resource.eval(IO.deferred[Set[RequestId]])
+            slowCoverageTarget    <- Resource.eval(IO.deferred[Set[Int]])
+            effectsLanded         <- Resource.eval(Ref[IO].of(Set.empty[TransactionHash]))
+            effectsLandedSignal   <- Resource.eval(IO.deferred[Unit])
+            effectsLandedTarget   <- Resource.eval(IO.deferred[List[BlockExpectation]])
+            fallbackEnteredSignal <- Resource.eval(IO.deferred[TransactionHash])
+            hooks = MultiPeerHeadHarness.Hooks[Stage4PeerHandle, Unit](
+                      peerTracer = peerNum =>
+                          Observers.captureStackHardConfirmed(
+                            peerNum,
+                            stacksMap,
+                            slowCoverageSignal,
+                            slowCoverageTarget,
+                          ) |+|
+                              Observers.captureBriefProduced(
+                                peerNum,
+                                blockBriefsMap,
+                                fastSettlementSignal,
+                                fastSettlementTarget,
+                              ) |+|
+                              Observers.captureTxSubmitting(
+                                effectsLanded,
+                                effectsLandedSignal,
+                                effectsLandedTarget,
+                              ) |+|
+                              Observers.captureFallbackEntered(fallbackEnteredSignal),
+                      coilTracer = coilNum =>
+                          Observers.captureCoilStackHardConfirmed(coilStacksMap(coilNum)) |+|
+                              Observers.captureTxSubmitting(
+                                effectsLanded,
+                                effectsLandedSignal,
+                                effectsLandedTarget,
+                              ) |+|
+                              Observers.captureFallbackEntered(fallbackEnteredSignal),
+                      peerHandle = (peerNum, conns) =>
+                          IO.pure(
+                            Stage4PeerHandle(
+                              conns.requestSequencer.getOrElse(
+                                sys.error(s"head peer $peerNum missing RequestSequencer")
                               )
                             )
-                        else
-                            val sleepMs = t.toEpochMilli - now.toEpochMilli
-                            val tickMs  = 5_000L
-                            val ticks   = sleepMs / tickMs
-                            val remMs   = sleepMs % tickMs
-                            log.info(s"WS mode: sleeping ${sleepMs / 1000}s until takeoff") >>
-                                (0L until ticks).toList.traverse_ { i =>
-                                    IO.sleep(tickMs.millis) >>
-                                        log.info(
-                                          s"WS takeoff in ${(sleepMs - (i + 1) * tickMs) / 1000}s"
-                                        )
-                                } >>
-                                IO.sleep(remMs.millis)
-                    }
-            }
-        } yield ()
-
-        // ------ Post-system, pre-stack IO: backend + per-peer Deferred/Ref maps. ------
-        val postSystem: IO[PostSystemState] = {
-            // All peers share one mock L1 backend, starting from the merged pre-init UTxOs.
-            // The head initialization tx is submitted by the protocol through normal operation.
-            val genesisUtxos = state.preinitPeerUtxosL1.values.reduce(_ ++ _)
-            for {
-                cardanoBackend <- CardanoBackendMock.mockIO(
-                  initialState = MockState(genesisUtxos),
-                  mkContext = slot =>
-                      Context(
-                        env = UtxoEnv(
-                          slot = slot,
-                          params = cardanoInfo.protocolParams,
-                          certState = CertState.empty,
-                          network = cardanoInfo.network
-                        ),
-                        slotConfig = cardanoInfo.slotConfig
-                      )
-                )
-                // Each peer gets its own PendingConnections deferred, completed after all actors
-                // are started so cross-peer liaisons can be wired.
-                pendingConnsMap <- peers
-                    .traverse { peerNum =>
-                        Deferred[IO, HeadMultisigRegimeManager.Connections].map(peerNum -> _)
-                    }
-                    .map(_.toMap)
-                // Per-peer Ref capturing every BlockBrief.Intermediate the peer's JointLedger emits.
-                // Populated via a ContraTracer sink attached to JL; consumed by propLiveness /
-                // propDepositTiming / propValidRatio / propStackCoverage.
-                blockBriefsMap <- peers
-                    .traverse { peerNum =>
-                        Ref[IO].of(Vector.empty[BlockBrief.Intermediate]).map(peerNum -> _)
-                    }
-                    .map(_.toMap)
-                stacksMap <- peers
-                    .traverse { peerNum =>
-                        Ref[IO].of(Vector.empty[Stack.HardConfirmed]).map(peerNum -> _)
-                    }
-                    .map(_.toMap)
-                coilStacksMap <- coilConfigs
-                    .traverse { coilConfig =>
-                        val coilNum = coilConfig.ownPeerId match {
-                            case PeerId.Coil(n) => n
-                            case PeerId.Head(_) =>
-                                throw new IllegalStateException(
-                                  "coil node config carries a head peer id"
-                                )
-                        }
-                        Ref[IO].of(Vector.empty[Stack.HardConfirmed]).map(coilNum -> _)
-                    }
-                    .map(_.toMap)
-                submittedRequestIds  <- Ref[IO].of(Vector.empty[RequestId])
-                fastSettlementSignal <- IO.deferred[Unit]
-                slowCoverageSignal   <- IO.deferred[Unit]
-                fastSettlementTarget <- IO.deferred[Set[RequestId]]
-                slowCoverageTarget   <- IO.deferred[Set[Int]]
-                effectsLanded        <- Ref[IO].of(Set.empty[TransactionHash])
-                effectsLandedSignal  <- IO.deferred[Unit]
-                effectsLandedTarget  <- IO.deferred[List[BlockExpectation]]
-                fallbackEnteredSignal <- IO.deferred[TransactionHash]
-            } yield PostSystemState(
-              cardanoBackend = cardanoBackend,
-              pendingConnsMap = pendingConnsMap,
-              blockBriefsMap = blockBriefsMap,
-              stacksMap = stacksMap,
-              coilStacksMap = coilStacksMap,
-              submittedRequestIds = submittedRequestIds,
-              fastSettlementSignal = fastSettlementSignal,
-              slowCoverageSignal = slowCoverageSignal,
-              fastSettlementTarget = fastSettlementTarget,
-              slowCoverageTarget = slowCoverageTarget,
-              effectsLanded = effectsLanded,
-              effectsLandedSignal = effectsLandedSignal,
-              effectsLandedTarget = effectsLandedTarget,
-              fallbackEnteredSignal = fallbackEnteredSignal,
-            )
-        }
-
-        case class PeerMrm(
-            mrm: HeadMultisigRegimeManager,
-            backendStore: BackendStore[IO],
+                          ),
+                      coilHandle = (_, _) => IO.unit,
+                    )
+            harness <- MultiPeerHeadHarness.resource(
+                           MultiPeerHeadHarness.Inputs(
+                             config = MultiPeerHeadHarness
+                                 .Config(label, backendMode, transportMode),
+                             multiNodeConfig = multiNodeConfig,
+                             coilNodeConfigs = coilConfigs,
+                             preinitPeerUtxosL1 = state.preinitPeerUtxosL1,
+                             takeoffTime = state.takeoffTime,
+                             startEpochMs = startEpochMs,
+                           ),
+                           hooks,
+                       )
+        yield Stage4Sut(
+          static = Stage4SutStatic(
+            system = harness.system,
+            cardanoBackend = harness.cardanoBackend,
+            peers = harness.peers.map { case (n, p) => n -> p.handle },
+            backendStores = harness.peers.map { case (n, p) => n -> p.backendStore },
+            log = Slf4jTracer.sink.contramap(Slf4jMsgFormat.humanFormat("Stage4.Sut")),
+          ),
+          mutable = Stage4SutMutable(
+            sutErrors = harness.sutErrors,
+            blockBriefs = blockBriefsMap,
+            stacks = stacksMap,
+            coilStacks = coilStacksMap,
+            submittedRequestIds = submittedRequestIds,
+            fastSettlementSignal = fastSettlementSignal,
+            slowCoverageSignal = slowCoverageSignal,
+            fastSettlementTarget = fastSettlementTarget,
+            slowCoverageTarget = slowCoverageTarget,
+            effectsLanded = effectsLanded,
+            effectsLandedSignal = effectsLandedSignal,
+            effectsLandedTarget = effectsLandedTarget,
+            fallbackEnteredSignal = fallbackEnteredSignal,
+          ),
         )
 
-        def buildPeerMrm(
-            peerNum: HeadPeerNumber,
-            system: ActorSystem[IO],
-            cardanoBackend: CardanoBackend[IO],
-            peerTransport: Resource[
-              IO,
-              ActorContext[IO, HeadMultisigRegimeManager.Request, Any] => PeerTransport
-            ],
-            hubCoilTransport: Option[Resource[
-              IO,
-              ActorContext[IO, HeadMultisigRegimeManager.Request, Any] => HubTransport
-            ]],
-            blockBriefsMap: Map[HeadPeerNumber, Ref[IO, Vector[BlockBrief.Intermediate]]],
-            stacksMap: Map[HeadPeerNumber, Ref[IO, Vector[Stack.HardConfirmed]]],
-            fastSettlementSignal: Deferred[IO, Unit],
-            slowCoverageSignal: Deferred[IO, Unit],
-            fastSettlementTarget: Deferred[IO, Set[RequestId]],
-            slowCoverageTarget: Deferred[IO, Set[Int]],
-            effectsLanded: Ref[IO, Set[TransactionHash]],
-            effectsLandedSignal: Deferred[IO, Unit],
-            effectsLandedTarget: Deferred[IO, List[BlockExpectation]],
-            fallbackEnteredSignal: Deferred[IO, TransactionHash],
-        ): Resource[IO, PeerMrm] = {
-            val nodeConfig = multiNodeConfig.nodeConfigs(peerNum)
-
-            // HMRM-level tracer = slf4j sink + per-peer capture observers (SCA stacks + JL briefs +
-            // CL TxSubmitting). Single source of truth feeds both the logger and the test's
-            // assertion refs.
-            val slf4jMrm: ContraTracer[IO, HeadMultisigRegimeManagerEvent] =
-                Slf4jTracer.sink.contramap(
-                  HeadMultisigRegimeManagerEventFormat.humanFormat(peerNum)
-                )
-            val mrmTracer =
-                slf4jMrm |+|
-                    Observers.captureStackHardConfirmed(
-                      peerNum,
-                      stacksMap,
-                      slowCoverageSignal,
-                      slowCoverageTarget,
-                    ) |+|
-                    Observers.captureBriefProduced(
-                      peerNum,
-                      blockBriefsMap,
-                      fastSettlementSignal,
-                      fastSettlementTarget,
-                    ) |+|
-                    Observers.captureTxSubmitting(
-                      effectsLanded,
-                      effectsLandedSignal,
-                      effectsLandedTarget,
-                    ) |+|
-                    Observers.captureFallbackEntered(fallbackEnteredSignal)
-
-            val persistenceTracer = Slf4jTracer.sink.contramap(PersistenceEventFormat.humanFormat)
-
-            openPeerBackend(
-              peerNum,
-              Cf.mkAll(
-                headPeers = multiNodeConfig.headConfig.headPeerNums.toList,
-                coilPeers = multiNodeConfig.headConfig.coilPeers.coilPeerNumbers,
-                hubs = multiNodeConfig.headConfig.coilPeers.hubHeadPeerNumbers
-              ),
-              persistenceTracer
-            ).flatMap { backendStore =>
-                for {
-                    persistence <- Resource.eval {
-                        given CardanoNetwork.Section = nodeConfig
-                        Persistence.fromBackend(backendStore, persistenceTracer)
-                    }
-                    l2Ledger <- Resource.eval(EutxoL2Ledger(nodeConfig))
-                    mrm <- HeadMultisigRegimeManager.resource(
-                      nodeConfig,
-                      cardanoBackend,
-                      l2Ledger,
-                      persistence,
-                      mrmTracer,
-                      peerTransport,
-                      hubCoilTransport,
-                    )
-                    _ <- Resource.eval(system.actorOf(mrm, s"hmrm-$peerNum"))
-                } yield PeerMrm(mrm, backendStore)
-            }
-        }
-
-        def acquireSutFromMrms(
-            system: ActorSystem[IO],
-            cardanoBackend: CardanoBackend[IO],
-            peerMrms: Map[HeadPeerNumber, PeerMrm],
-            coilMrms: Map[CoilPeerNumber, CoilMrm],
-            blockBriefsMap: Map[HeadPeerNumber, Ref[IO, Vector[BlockBrief.Intermediate]]],
-            stacksMap: Map[HeadPeerNumber, Ref[IO, Vector[Stack.HardConfirmed]]],
-            coilStacksMap: Map[CoilPeerNumber, Ref[IO, Vector[Stack.HardConfirmed]]],
-            submittedRequestIds: Ref[IO, Vector[RequestId]],
-            fastSettlementSignal: Deferred[IO, Unit],
-            slowCoverageSignal: Deferred[IO, Unit],
-            fastSettlementTarget: Deferred[IO, Set[RequestId]],
-            slowCoverageTarget: Deferred[IO, Set[Int]],
-            effectsLanded: Ref[IO, Set[TransactionHash]],
-            effectsLandedSignal: Deferred[IO, Unit],
-            effectsLandedTarget: Deferred[IO, List[BlockExpectation]],
-            fallbackEnteredSignal: Deferred[IO, TransactionHash],
-        ): IO[Stage4Sut] =
-            for {
-                peerConnections <- peerMrms.toList
-                    .traverse { case (peerNum, peerMrm) =>
-                        peerMrm.mrm.connectionsDeferred.get.map(peerNum -> _)
-                    }
-                    .map(_.toMap)
-                coilConnections <- coilMrms.toList
-                    .traverse { case (coilNum, coilMrm) =>
-                        coilMrm.mrm.connectionsDeferred.get.map(coilNum -> _)
-                    }
-                    .map(_.toMap)
-                sutErrors <- Ref[IO].of(List.empty[String])
-                errorDrainer <- system.eventStream.take
-                    .flatMap {
-                        case e: ActorError if e.cause != ActorError.NoCause =>
-                            sutErrors.update(_ :+ s"[${e.logSource}] ${e.cause.getMessage}")
-                        case _ => IO.unit
-                    }
-                    .foreverM
-                    .start
-                headTickFibers <- peerConnections.toList.traverse { case (peerNum, conns) =>
-                    val pollingPeriod = multiNodeConfig
-                        .nodeConfigs(peerNum)
-                        .nodeOperationMultisigConfig
-                        .cardanoLiaisonPollingPeriod
-                    (IO.sleep(pollingPeriod) >>
-                        (conns.cardanoLiaison ! CardanoLiaison.Timeout)).foreverM.start
-                }
-                coilTickFibers <- coilConnections.toList.traverse { case (coilNum, conns) =>
-                    val pollingPeriod = coilMrms(coilNum).config
-                        .nodeOperationMultisigConfig
-                        .cardanoLiaisonPollingPeriod
-                    (IO.sleep(pollingPeriod) >>
-                        (conns.cardanoLiaison ! CardanoLiaison.Timeout)).foreverM.start
-                }
-            } yield Stage4Sut(
-              static = Stage4SutStatic(
-                system = system,
-                cardanoBackend = cardanoBackend,
-                peers = peerConnections.map { case (peerNum, conns) =>
-                    peerNum -> Stage4PeerHandle(
-                      conns.requestSequencer.getOrElse(
-                        sys.error(s"head peer $peerNum missing RequestSequencer")
-                      )
-                    )
-                },
-                errorDrainer = errorDrainer,
-                liaisonTickFibers = headTickFibers ++ coilTickFibers,
-                backendStores = peerMrms.map { case (peerNum, peerMrm) =>
-                    peerNum -> peerMrm.backendStore
-                },
-                log = Slf4jTracer.sink.contramap(Slf4jMsgFormat.humanFormat("Stage4.Sut")),
-              ),
-              mutable = Stage4SutMutable(
-                sutErrors = sutErrors,
-                blockBriefs = blockBriefsMap,
-                stacks = stacksMap,
-                coilStacks = coilStacksMap,
-                submittedRequestIds = submittedRequestIds,
-                fastSettlementSignal = fastSettlementSignal,
-                slowCoverageSignal = slowCoverageSignal,
-                fastSettlementTarget = fastSettlementTarget,
-                slowCoverageTarget = slowCoverageTarget,
-                effectsLanded = effectsLanded,
-                effectsLandedSignal = effectsLandedSignal,
-                effectsLandedTarget = effectsLandedTarget,
-                fallbackEnteredSignal = fallbackEnteredSignal,
-              ),
-            )
-
-        case class CoilMrm(
-            mrm: CoilMultisigRegimeManager,
-            backendStore: BackendStore[IO],
-            config: NodeConfig,
-        )
-
-        /** Concrete WS handle + the OS-assigned port the NodeWsServer bound to. Carried in
-          * [[HeadNetwork.ws]] so the outer for-comprehension can compute remote URIs from real
-          * bound ports (post-bind) and start every peer's dialers in a second pass.
-          */
-        case class WsBinding(
-            wsPeerTransport: WsPeerTransport,
-            boundPort: Int,
-        )
-
-        case class HeadNetwork(
-            peerTransport: PeerTransport,
-            hubTransport: Option[HubTransport],
-            ws: Option[WsBinding],
-        )
-
-        // Per-head-peer transports: head-mesh PeerTransport (always) plus, for hub head peers, the
-        // hub-side HubTransport for the hub↔coil link. WS mode mounts both routes on ONE shared
-        // NodeWsServer per peer (bound on port 0 — OS-assigned ephemeral — so two WS test instances
-        // never collide); Direct mode wires them to the shared in-process registries instead.
-        def headNetworkFor(
-            peerNum: HeadPeerNumber,
-            inProcessRegistry: InProcessPeerTransport.Registry,
-            hubCoilRegistry: Option[InProcessHubCoilTransport.Registry],
-        ): Resource[IO, HeadNetwork] = {
-            val ownHeadPeerId = headPeerId(multiNodeConfig, peerNum)
-            val hubbedCoils = multiNodeConfig.headConfig.hubbedCoilPeerNums(peerNum)
-            transportMode match {
-                case TransportMode.Direct =>
-                    for {
-                        peerT <- Resource.eval(
-                          InProcessPeerTransport.create(ownHeadPeerId, inProcessRegistry)
-                        )
-                        hubT <-
-                            if hubbedCoils.isEmpty then
-                                Resource.pure[IO, Option[HubTransport]](None)
-                            else
-                                hubCoilRegistry match {
-                                    case None =>
-                                        Resource.eval(
-                                          IO.raiseError(
-                                            new IllegalStateException(
-                                              s"head peer $peerNum hubs coil peers in Direct mode " +
-                                                  "but no hubCoilRegistry was allocated"
-                                            )
-                                          )
-                                        )
-                                    case Some(reg) =>
-                                        Resource
-                                            .eval(InProcessHubCoilTransport.Hub.create(reg))
-                                            .map(h => Some(h: HubTransport))
-                                }
-                    } yield HeadNetwork(peerT, hubT, None)
-                case TransportMode.WebSocket =>
-                    given CardanoNetwork.Section = multiNodeConfig.headConfig
-                    val remoteIds: List[HeadPeerId] =
-                        peers.filterNot(_ == peerNum).map(headPeerId(multiNodeConfig, _)).toList
-                    val ptTracer =
-                        Slf4jTracer.sink.contramap(PeerTransportEventFormat.humanFormat(peerNum))
-                    val nwsTracer =
-                        Slf4jTracer.sink.contramap(NodeWsServerEventFormat.humanFormat(peerNum))
-                    val hubTracer =
-                        Slf4jTracer.sink.contramap(HubWsTransportEventFormat.humanFormat(peerNum))
-                    val bindHost = host"127.0.0.1"
-                    for {
-                        peerT <- Resource.eval(
-                          WsPeerTransport.create(ownHeadPeerId, remoteIds, ptTracer)
-                        )
-                        hubTConcrete: Option[HubWsTransport] <-
-                            if hubbedCoils.isEmpty then
-                                Resource.pure[IO, Option[HubWsTransport]](None)
-                            else
-                                Resource
-                                    .eval(HubWsTransport.create(hubbedCoils, hubTracer))
-                                    .map(Some(_))
-                        meshRoute = (wsb: WebSocketBuilder2[IO]) => peerT.routes(wsb)
-                        hubRoutes = hubTConcrete.toList.map(h =>
-                            (wsb: WebSocketBuilder2[IO]) => h.routes(wsb)
-                        )
-                        server <- NodeWsServer.resource(
-                          bindHost,
-                          Port.fromInt(0).get,
-                          meshRoute :: hubRoutes,
-                          nwsTracer,
-                        )
-                        boundPort = server.address.getPort
-                    } yield HeadNetwork(
-                      peerT,
-                      hubTConcrete.map(h => h: HubTransport),
-                      Some(WsBinding(peerT, boundPort)),
-                    )
-            }
-        }
-
-        // Per-coil-peer uplink transport toward its single hub. Direct mode uses the shared
-        // in-process hub-coil registry; WS mode dials the hub's `/hub` route at the hub's
-        // post-bind ephemeral port (passed in via `hubBoundPort`).
-        def coilUplinkFor(
-            coilNum: CoilPeerNumber,
-            hubCoilRegistry: InProcessHubCoilTransport.Registry,
-            wsClient: Option[WSClient[IO]],
-            hubBoundPort: Option[Int],
-        ): Resource[
-          IO,
-          ActorContext[IO, HeadMultisigRegimeManager.Request, Any] => CoilTransport
-        ] = transportMode match {
-            case TransportMode.Direct =>
-                Resource
-                    .eval(InProcessHubCoilTransport.Coil.create(coilNum, hubCoilRegistry))
-                    .map(t =>
-                        (_: ActorContext[IO, HeadMultisigRegimeManager.Request, Any]) =>
-                            t: CoilTransport
-                    )
-            case TransportMode.WebSocket =>
-                given CardanoNetwork.Section = multiNodeConfig.headConfig
-                val port = hubBoundPort.getOrElse(
-                  sys.error(s"coil $coilNum's hub has no bound port — head networks not built yet?")
-                )
-                val hubUri = Uri.unsafeFromString(s"ws://127.0.0.1:$port/hub")
-                val cpwtTracer =
-                    Slf4jTracer.sink.contramap(CoilPeerWsTransportEventFormat.humanFormat(coilNum))
-                val client = wsClient.getOrElse(
-                  sys.error("WebSocket transport requires a shared JdkWSClient")
-                )
-                for {
-                    t <- Resource.eval(CoilPeerWsTransport.create(coilNum, cpwtTracer))
-                    _ <- t.startDialer(client, hubUri)
-                } yield (_: ActorContext[IO, HeadMultisigRegimeManager.Request, Any]) =>
-                    t: CoilTransport
-        }
-
-        // ------ Per-coil-peer follower stack: full CoilMultisigRegimeManager + its own in-memory
-        // ------ backend + persistence + L2 ledger, with the SCA capture sink wired into the
-        // ------ MRM tracer so hard-confirmed stacks land in `coilStacksRef`.
-        def buildCoilMrm(
-            coilConfig: NodeConfig,
-            coilNum: CoilPeerNumber,
-            system: ActorSystem[IO],
-            cardanoBackend: CardanoBackend[IO],
-            coilTransport: Resource[
-              IO,
-              ActorContext[IO, HeadMultisigRegimeManager.Request, Any] => CoilTransport
-            ],
-            coilStacksRef: Ref[IO, Vector[Stack.HardConfirmed]],
-            effectsLanded: Ref[IO, Set[TransactionHash]],
-            effectsLandedSignal: Deferred[IO, Unit],
-            effectsLandedTarget: Deferred[IO, List[BlockExpectation]],
-            fallbackEnteredSignal: Deferred[IO, TransactionHash],
-        ): Resource[IO, CoilMrm] = {
-            // Reuse the head-typed event format with a synthetic peer number (matches the legacy
-            // labelling); a coil-specific format is a separate follow-up.
-            val labelNum = HeadPeerNumber(nPeers + coilNum.convert)
-            val slf4jMrm: ContraTracer[IO, HeadMultisigRegimeManagerEvent] =
-                Slf4jTracer.sink.contramap(
-                  HeadMultisigRegimeManagerEventFormat.humanFormat(labelNum)
-                )
-            // Coil also runs a CardanoLiaison and can submit backbone txs to L1 (e.g. the init tx
-            // in some races), so the TxSubmitting + FallbackToRuleBased capture observers must be
-            // wired to its tracer too — otherwise the test's landed set and fallback signal miss
-            // coil-side submissions.
-            val mrmTracer =
-                slf4jMrm |+|
-                    Observers.captureCoilStackHardConfirmed(coilStacksRef) |+|
-                    Observers.captureTxSubmitting(
-                      effectsLanded,
-                      effectsLandedSignal,
-                      effectsLandedTarget,
-                    ) |+|
-                    Observers.captureFallbackEntered(fallbackEnteredSignal)
-            val persistenceTracer = Slf4jTracer.sink.contramap(PersistenceEventFormat.humanFormat)
-
-            InMemoryBackendStore.open(persistenceTracer).flatMap { backendStore =>
-                for {
-                    persistence <- Resource.eval {
-                        given CardanoNetwork.Section = coilConfig
-                        Persistence.fromBackend(backendStore, persistenceTracer)
-                    }
-                    l2Ledger <- Resource.eval(EutxoL2Ledger(coilConfig))
-                    mrm <- CoilMultisigRegimeManager.resource(
-                      coilConfig,
-                      cardanoBackend,
-                      l2Ledger,
-                      persistence,
-                      mrmTracer,
-                      coilTransport,
-                    )
-                    _ <- Resource.eval(system.actorOf(mrm, s"cmrm-${coilNum.convert}"))
-                } yield CoilMrm(mrm, backendStore, coilConfig)
-            }
-        }
-
-        for {
-            _      <- Resource.eval(preSystem)
-            system <- ActorSystem[IO](label)
-            pss    <- Resource.eval(postSystem)
-            inProcessRegistry <- Resource.eval(InProcessPeerTransport.emptyRegistry)
-            // Allocated only when this scenario has coil peers; in Direct mode it backs the
-            // in-process hub↔coil routing, and in WS mode the field stays `None` (the hub/coil
-            // WS transports own their own routing). Heads with no hubbed coils pass `None`
-            // either way.
-            hubCoilRegistry <-
-                if coilConfigs.isEmpty then
-                    Resource.pure[IO, Option[InProcessHubCoilTransport.Registry]](None)
-                else
-                    Resource
-                        .eval(InProcessHubCoilTransport.emptyRegistry)
-                        .map(Some(_))
-            // Shared WS client: allocated once for WebSocket mode, unused for Direct.
-            wsClient <- transportMode match {
-                case TransportMode.Direct =>
-                    Resource.pure[IO, Option[WSClient[IO]]](None)
-                case TransportMode.WebSocket =>
-                    Resource.eval(JdkWSClient.simple[IO]).map(Some(_))
-            }
-            // Pass 1: allocate every head peer's transports + bind their NodeWsServers (WS mode on
-            // port 0 ⇒ OS-assigned ephemeral). Each `HeadNetwork` carries the bound port; dialers
-            // are deferred to Pass 2 so URIs can be built from the real port map.
-            headNetworks <- peers.toList
-                .traverse { peerNum =>
-                    headNetworkFor(peerNum, inProcessRegistry, hubCoilRegistry)
-                        .map(peerNum -> _)
-                }
-                .map(_.toMap)
-            // Pass 2 (WS only): now that every peer's server is bound and we know its ephemeral
-            // port, build the `HeadPeerId -> Uri` map and start each peer's dialers against the
-            // ones with higher peerNum (lower dials higher; matches the production wiring).
-            _ <- transportMode match {
-                case TransportMode.Direct => Resource.pure[IO, Unit](())
-                case TransportMode.WebSocket =>
-                    val client = wsClient.getOrElse(
-                      sys.error("WebSocket transport requires a shared JdkWSClient")
-                    )
-                    val peerHeadUris: Map[HeadPeerId, Uri] = peers.map { p =>
-                        val port = headNetworks(p).ws
-                            .map(_.boundPort)
-                            .getOrElse(sys.error(s"peer $p has no WS binding in WS mode"))
-                        headPeerId(multiNodeConfig, p) -> Uri.unsafeFromString(
-                          s"ws://127.0.0.1:$port/head"
-                        )
-                    }.toMap
-                    peers.toList.traverse_ { peerNum =>
-                        val ownId = headPeerId(multiNodeConfig, peerNum)
-                        val ws = headNetworks(peerNum).ws.get
-                        ws.wsPeerTransport.startDialers(client, peerHeadUris - ownId)
-                    }
-            }
-            peerMrms <- peers.toList
-                .traverse { peerNum =>
-                    val net = headNetworks(peerNum)
-                    val peerFactory: Resource[
-                      IO,
-                      ActorContext[IO, HeadMultisigRegimeManager.Request, Any] => PeerTransport
-                    ] =
-                        Resource.pure(
-                          (_: ActorContext[IO, HeadMultisigRegimeManager.Request, Any]) =>
-                              net.peerTransport
-                        )
-                    val hubFactory: Option[Resource[
-                      IO,
-                      ActorContext[IO, HeadMultisigRegimeManager.Request, Any] => HubTransport
-                    ]] = net.hubTransport.map { h =>
-                        Resource.pure(
-                          (_: ActorContext[IO, HeadMultisigRegimeManager.Request, Any]) => h
-                        )
-                    }
-                    buildPeerMrm(
-                      peerNum,
-                      system,
-                      pss.cardanoBackend,
-                      peerFactory,
-                      hubFactory,
-                      pss.blockBriefsMap,
-                      pss.stacksMap,
-                      pss.fastSettlementSignal,
-                      pss.slowCoverageSignal,
-                      pss.fastSettlementTarget,
-                      pss.slowCoverageTarget,
-                      pss.effectsLanded,
-                      pss.effectsLandedSignal,
-                      pss.effectsLandedTarget,
-                      pss.fallbackEnteredSignal,
-                    ).map(peerNum -> _)
-                }
-                .map(_.toMap)
-            // Per-coil follower stack. The MRM gets its uplink transport via `coilUplinkFor`; in
-            // Direct mode this shares `hubCoilRegistry` with the hub head peer's `HubTransport`.
-            coilMrms <- coilConfigs
-                .traverse { coilConfig =>
-                    val coilNum = coilConfig.ownPeerId match {
-                        case PeerId.Coil(n) => n
-                        case PeerId.Head(_) =>
-                            throw new IllegalStateException(
-                              "coil node config carries a head peer id"
-                            )
-                    }
-                    val hubNum = coilConfig
-                        .coilPeerHub(coilNum)
-                        .getOrElse(
-                          throw new IllegalStateException(
-                            s"no hub configured for coil peer $coilNum"
-                          )
-                        )
-                    val registry = hubCoilRegistry.getOrElse(
-                      throw new IllegalStateException(
-                        "coilConfigs is non-empty but hubCoilRegistry was not allocated"
-                      )
-                    )
-                    val hubBoundPort = headNetworks(hubNum).ws.map(_.boundPort)
-                    buildCoilMrm(
-                      coilConfig,
-                      coilNum,
-                      system,
-                      pss.cardanoBackend,
-                      coilUplinkFor(coilNum, registry, wsClient, hubBoundPort),
-                      pss.coilStacksMap(coilNum),
-                      pss.effectsLanded,
-                      pss.effectsLandedSignal,
-                      pss.effectsLandedTarget,
-                      pss.fallbackEnteredSignal,
-                    ).map(coilNum -> _)
-                }
-                .map(_.toMap)
-            sut <- Resource.make(
-              acquireSutFromMrms(
-                system,
-                pss.cardanoBackend,
-                peerMrms,
-                coilMrms,
-                pss.blockBriefsMap,
-                pss.stacksMap,
-                pss.coilStacksMap,
-                pss.submittedRequestIds,
-                pss.fastSettlementSignal,
-                pss.slowCoverageSignal,
-                pss.fastSettlementTarget,
-                pss.slowCoverageTarget,
-                pss.effectsLanded,
-                pss.effectsLandedSignal,
-                pss.effectsLandedTarget,
-                pss.fallbackEnteredSignal,
-              )
-            )(sut =>
-                sut.static.liaisonTickFibers.traverse_(_.cancel) >>
-                    IO.sleep(100.millis) >>
-                    sut.static.errorDrainer.cancel
-            )
-        } yield sut
-    }
-
-    /** This peer's [[HeadPeerId]] — derived from its number + the head-set size; `NodeConfig` does
-      * not expose a head-specific id.
-      */
-    private def headPeerId(
-        multiNodeConfig: MultiNodeConfig,
-        peerNum: HeadPeerNumber
-    ): HeadPeerId =
-        HeadPeerId(peerNum, multiNodeConfig.nHeadPeers)
 
     override def beforeFinalize(lastState: ModelState, sut: Stage4Sut): IO[Prop] = {
         // Race the happy-path drain against the fallback-entered signal. If any peer's CL
@@ -1127,25 +510,6 @@ case class Stage4Suite(
             }
             .map(_.foldLeft(Prop.passed)(_ && _))
     }
-
-    /** Per-peer backend allocator chosen by [[backendMode]].
-      *
-      * `BackendMode.InMemory` returns a fresh `InMemoryBackendStore`. `BackendMode.RocksDb(root)`
-      * opens `root/peer-N/`, creating parent directories on demand. The returned `Resource` is
-      * allocated immediately in `sutResource`; cleanup currently leaks RocksDB handles until a
-      * stage4-level shutdown hook lands (parallel to `errorDrainer.cancel`).
-      */
-    private def openPeerBackend(
-        peerNum: HeadPeerNumber,
-        cfs: List[Cf],
-        tracer: ContraTracer[IO, PersistenceEvent]
-    ): Resource[IO, BackendStore[IO]] =
-        backendMode match
-            case BackendMode.InMemory => InMemoryBackendStore.open(tracer)
-            case BackendMode.RocksDb(root) =>
-                val dir = root.resolve(s"peer-${peerNum: Int}")
-                Resource.eval(IO.blocking(Files.createDirectories(dir))) >>
-                    RocksDbBackendStore.open(dir, cfs, tracer)
 
     private def countEntries(
         backend: BackendStore[IO],
